+++ src/modes/select/select.mode.ts
@@ -226,6 +226,7 @@ export class TerraDrawSelectMode extends TerraDrawBaseSelectMode<SelectionStylin
 			this.pixelDistance,
 			this.selectionPoints,
 			this.midPoints,
+			10,
 		);
 	}
 
@@ -410,7 +411,7 @@ export class TerraDrawSelectMode extends TerraDrawBaseSelectMode<SelectionStylin
 
 		// Select feature
 		this.selected = [featureId];
-		console.log("setting selected");
+
 		this.store.updateProperty([
 			{ id: featureId, property: "selected", value: true },
 		]);